io.ItemState as interface #1053
